Added : Changed backend functions to take markdown data and upload it, made a events details page , made loading inside the layout so the navbar is rendered

+++ src/lib/supabase/actions/activities.actions.ts
@@ -1,5 +1,6 @@
 import { FormActivityType } from "@/types";
 import { client } from "../supabase"
+import { deleteMarkdownFile, deleteMarkdownFolder, uploadMarkdownFile } from "./storage.actions";
 
 export const getActivites = async () => {
     const { data, error } = await client.from("activities").select("*");
@@ -15,9 +16,23 @@ export const getActivityById = async (id: string) => {
     return JSON.parse(JSON.stringify(data[0]));
 };
 
+export const updateActivity = async (activity: FormActivityType) => {
+    const { id, longDescription, ...rest } = activity;
+    await deleteMarkdownFile(`${id}.md`, "activities");
+    await uploadMarkdownFile(`${id}.md`, "activities", longDescription);
+    const { error } = await client.from("activities").update(rest).eq("id", activity.id);
+    if (error) {
+        console.log(error);
+    }
+    return error;
+};
+
 export const uploadActivity = async (activity: FormActivityType) => {
-    const { id, ...rest } = activity;
-    const { error } = await client.from("activities").insert(rest);
+    // upload the activity -> upload the markdown file with the name === id
+    const { id, longDescription, ...rest } = activity;
+    const { data, error } = await client.from("activities").insert(rest).select().single();
+    await uploadMarkdownFile(`${data.id}.md`, "activities", longDescription);
+
     if (error) {
         console.log(error);
         return { error: error };
@@ -26,16 +41,16 @@ export const uploadActivity = async (activity: FormActivityType) => {
 
 };
 
-export const updateActivity = async (activity: FormActivityType) => {
-    const { id, ...rest } = activity;
-    const { error } = await client.from("activities").update(rest).eq("id", activity.id);
-    if (error) {
-        console.log(error);
-    }
-    return error;
-};
+
 
 export const deleteActivity = async (id: string) => {
+    const data = await deleteMarkdownFolder(id, "activities");
+
+    if (!data) {
+        throw new Error(`Markdown file of ${id} in activities folder could not be deleted`);
+    }
+
     const response = await client.from("activities").delete().eq("id", id);
+
     return response;
 };
